+++ spec/spec_helper.rb
@@ -33,3 +33,14 @@ def build_interactor(*modules, &block)
   interactor.class_eval(&block) if block
   interactor
 end
+
+def expect_piping_error_called(message_substring:, error_class:, error_message:, action: nil)
+  matcher = {
+    exception: an_object_satisfying { |e| e.is_a?(error_class) && e.message == error_message },
+  }
+  matcher[:action] = action unless action.nil?
+  expect(Axn::Util).to have_received(:piping_error).with(
+    a_string_including(message_substring),
+    hash_including(matcher),
+  )
+end